+++ frontend/src/components/AnnotationView/CardDataset.vue
@@ -1,12 +1,12 @@
 <template>
-    <div class="model-cards">
-      <!-- TODO: This is intended to become a <slot> to allow any card type to be used here -->
-      <AllianceGenomeCard
-        v-for="(model) in this.models"
-        :key="model.id"
-        :model="model"
-      />
-    </div>
+  <div class="model-cards">
+    <!-- TODO: This is intended to become a <slot> to allow any card type to be used here -->
+    <AllianceGenomeCard
+      v-for="(model) in this.models"
+      :key="model.id"
+      :model="model"
+    />
+  </div>
 </template>
 
 <script>
@@ -39,8 +39,6 @@ export default {
   display: flex;
   flex-wrap: wrap;
   gap: 1rem;
-  padding: 1rem;
-  height: 100%;
 }
 
 </style>